requirements.txt
fastmcp>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
uvicorn>=0.30.0
